(#7934) Give a better error message when filebucket contents don't exist
Before if the filebucket content didn't exist when you clicked the link
to it, it would raise an exception that we wouldn't catch, so in
production you'd get a generic error message.  Now it will tell you the
filebucket content doesn't exist and link to documentation on how to
enable the file bucket.

Also fixed some other bugs such as rendering text if you passed an
invalid md5 and improved test robustness.

Based on code from Jacob Helwig <jacob@puppetlabs.com>
Reviewed-by: Nick Lewis <nick@puppetlabs.com>