maint: Fix node_ids method by not overwriting it with and attr_accessor
Rails provides some magic methods for models that give you the ids of
the children for a model with a `child_ids` method.  We weren't using
this until recently, but when I added an accesor for node_ids in order
to get groups to assign nodes on creation, the accessor broke the Rails
_ids methods.

This change renames the accessors to use 'assigned' in the name so they
don't conflict with the rails methods.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>