maint: Move duplicated code to a helper method
The node and node_group both had nearly identical code to assign related
models, so we moved it to a Module that both models were already using.

Also, the view code was duplicated so we moved it to a partial.

We also made it so that assigning node groups didn't fail when node
classification was disabled, just fail when node classes are assigned.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>