(#9262) causes the contents of .section divs to scroll when overflowing
This is a short-term solution to a larger problem. Dashboard is split into two columns,
the sidebar (20% of window width), and the main content pane (80% of window width). The
problem is when dealing with large strings of non-breaking text (e.g. file paths), the
containing element runs off the screen unless accomodations are made (e.g. overflow: auto;).
The better--long term--fix is to replace the 20%/80% column layout with one that allows
the main content pane variable width. Thus, the entire window scrolls horizonally when
elements overflow instead of requiring short-term fixes for each offending element.