(#9423) Fix broken RH init script in dashboard
A typo in the chkconfig values for the puppet-dashboard-workers init script was
making it unusable. Adding the missing value will address the issue. This patch
also modifies the default start runlevel to exclude 2, as level 2 does not
start networking or network services. Runlevel 2 has been added to the default
stop level.

Signed-off-by: Matthaus Litteken <matthaus@puppetlabs.com>