(#9282) Rewrite explanatory text in database.yml.example
The inline documentation in the example database.yml file was not as clear as
it could be, and suffered from some redundancy. It also did not explain
requirements about when Dashboard requires its database/user to exist and did
not make a strong enough recommendation about which environment most users
should use.

This commit updates the inline documentation to more closely match
the current Dashboard manual as of version 1.2.