(#9756) Autoload plugin initializers.
Rails has a robust built-in mechanism for handling segregated
application-specific configuration code.  This initializer takes
the next step and uses the same approach to load initializers from
installed plugins.

This ends up being beneficial over the standard plugin `init.rb`
not only for code segregation reasons, but because `init.rb` is
run before any of the core initializers are.  Because Dashboard's
`SETTINGS` are loaded in the first initializer, this also gives us
the cleanest mechanism for writing initialization code that depends
upon or alters the `SETTINGS` object.