(#9756/Maint) Better guards around installed_plugins.
The initializer added by #9756 didn't play nicely with a
Dashboard that had no plugins installed, and the Registry
didn't place nicely when you dropped a plugin into place
but didn't have an `installed_plugins` directory.

As it turns out, `Dir#glob` does exactly the right thing
in the event that any element in the path doesn't exist,
allowing us to simplify the code while simultaneously
making it more robust.