(#9773) Report content hidden
If any cell in a report table had wide content, the part of the table that would
have extended beyond the viewable window was hidden.  A scroll bar was provided
at the bottom of the content, but it was hard to find.

This change allows the table expand to the size it needs to be to display all
its content.  It allows for a horizontal scroll bar on the browser window, which
is much more visible.