(#9366) allows a custom version link to be specified
When a 'VERSION_LINK' file is found in the root of the app, the
contents are read and used as a custom URL for the version link
in the header.

This behavior is similar to when a 'VERSION' file is found in
the root. If the file is missing--and the app is a git checkout,
a link to the current changelog version on github is constructed.