(#10024) Update Puppet Dashboard Workers init script
Minor fixes to dashboard-workers script. The script sources a sysconfig
file that uses DASHBOARD_HOME rather than DASHBOARD_ROOT, so this script
has been updated to utilize the DASHBOARD_HOME env variable.
Additionally the sysconfig file had a value for the user to run the
processes as, but then was hard-coding the user-name. Script has been
updated to allow changing of username, as per the sysconfig file.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>