(#6717) Add foreign keys and cascading deletes to report related tables
Using the `:dependent => :destroy` on Rails records is really slow when
you need to delete a lot of child records since it makes an active
record object for each one.  Instead, if we have foreign keys with
cascading deletes, this will happen very quickly.

We wouldn't want this if we were relying on hooks with destroy, but the
only model that has that kind of hook is for reports, and it just
updates the associated node table, but if you've deleted the node, that
hook will be useless.

Reviewed-by: Max Martin <max@puppetlabs.com>