(#6717) Update last report on the node after deleting reports
When calling delete instead of destroy on reports the callback to update
the node info doesn't get called.  Delete is a lot faster since you
don't need to create an ActiveRecord object to do it, so we'll leave it
as delete and manually do the callback to update nodes after we've taken
care of all the reports.

Reviewed-by: Max Martin <max@puppetlabs.com>