(#11063) Building deb package should keep the source tarball
When building the deb package, the source tarball does not get saved in
the pkg directory. Since this could be something useful to keep around,
we should save the file. We do this by copying the latest tarball out of
the temporary directory. The tar rake task also generates a tarball, but
we are guaranteed that the debian source tarball will have a later mtime
than that one.