(#4890) Remove AllowOverride AuthConfig from example vhost
Unless you consider its javascript and CSS files to be confidential,
there is absolutely no situation where you'd use an .htaccess file to override
auth for Dashboard, because access to locations in a Rails app can't be
controlled with a directory override. This commit kills a misleading example
line that implied this was possible.