(#4890) Add example configuration for using basic auth with Dashboard
Pending a fix for issue #7173, Puppet can't submit reports to a server that
requires authentication. A known way to get around this is to provide
exceptions in Dashboard's security configuration for retrieving node
definitions and submitting reports. This commit adds a commented-out section
to the example vhost file to do exactly that.