(#12440) Add config flag to puppet-dashboard file in rh spec
Without this commit, the installer for dashboard treats the
/etc/sysconfig/puppet-dashboard config file as a regular file
and overwrites it during install, even if it already exists
and has been modified.  This patch marks puppet-dashboard in
the spec file as a config file, so that if it exists on the
install target and is modified, the original version is preserved.