($12440) Add additional config file to spec file
Without this commit, the /etc/logrotate.d/puppet-dashboard
file is overwritten during install, even if it exists and
has been modified. This patch flags the logrotate.d/puppet-dashboard
file as a config file, so that if it exists on the install
target and is modified, the original version is preserved.

Signed-off-by: Moses Mendoza moses@puppetlabs.com