(#12440) change config behavior for logrotate.d/puppet-dashboard
This patch modifies the rpm behavior with respect to
logrotate.d/puppet-dashboard, marking it as a standard config
file without the noreplace tag. Without this modification,
if a puppet install were to require install of a new
puppet-dashboard logrotate file, any user mods of the previous
version would prevent the installation, instead causing the new
config file to be installed as <config>.rpmnew.

This patch changes the behavior to instead install the new
config file by default, and save the previous file (if edited)
as <config>.rpmsave.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>