Add parameters support for nodes and node groups
Previous to this commit there was no rake API for setting parameters on
nodes and node groups.  This commit adds the `node:parameters` and
`nodegroup:parameters` tasks. Both accept two arguments:
- *name* Which node or node group to apply the parameters to
- *parameters* In format param1=val,param2=val2

This format was chosen for two reasons. I wanted multiple parameters to
be
able to supported on a single command. Further, rake doesn't support
multiple instances of the same argument so a format such as
'parameter=param,value parameter=param2,value2' could not be supported.

Example:
`rake node:parameters name=master.puppet.internal \
parameters=ntpserver=ntp1.mydomain.com,dnsserver=dns.mydomain.com`

Potential Issues:
Currently it is impossible to escape the *=* or *,* characters when
specifying the parameters value.  So node or node group parameters that
include a comma or equal sign will not be able to be specified.