Add VERSION file to source
In dashboard packaging we currently just echo
a version and redirect to a file within dashboard.
This commit adds it to the source so we can
automate its incrementing with the packaging repo
and ditch the wierd automation we had set up to
work around its non-existence in the git tree.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>