(#16333) Override AR MySQL connector to set sql_mode
Currently if there is a problematic sql_mode set globally it will end up
breaking things. This patch overrides the configure_connection function
of the MySQL connector to explicitly set the sql_mode to traditional for
the session so that if something that breaks the queries used by the
console (like ANSI_QUOTES) is set globally, it won't affect it.