(#9296) Fix loading rack for rubygems >= 1.8
Rubygems version 1.8 broke some of the ways vendored gems used to work in Rails
2. This commit updates the config/environment.rb to add all of the gem libs to
the rails autoload_paths. It also updates the actionpack action_controller
library to not call out to gem explicitly. We know that the correct rack
version is being supplied in the vendored gems, so the following require 'rack'
line will succeed.

This also munges the load path in config/boot.rb because boot.rb doesn't load
environment.rb right away, and so can't find rack, due to the same gem problem.
So this uses the same loop as in environment.rb, but instead of the rails
autoload_paths, it adds the libs to $:, or $LOAD_PATH.