(#PE-339) Initialize pe_version to automate fact
Facter can be an inconsistent way to pass information, and can also
include a large overhead. This commit initializes a fact that is used in
several repos, the current major.minor pe_version.  This provides a
simple, centralized way to access this versioning information, which
also automates it when Facter facts are updated. The fact is read from a
file (/opt/puppet/pe_version), so changes to that path will require a
change in this initializer, pe_version_init.rb.