Add listclasses, delclass to nodegroup rake tasks
Previously once a class had been added to a nodegroup, there was no way of
removing it, short of deleting the group and adding the desired classes back.
However, if you wanted to list the classes in that group, you were still out of
luck, because there was no listclasses task for the nodegroup namespace. This
commit adds both of those tasks to the nodegroup namespace. listclasses lists
the classes assigned to a nodegroup and delclass removes a class from a
nodegroup, if such a class exists in the nodegroup.