Add addgroup, addclass, listgroups, listclasses tasks for nodes
Previously if a user wanted to use rake tasks to interrogate the dashboard
about which groups or classes a node belonged to, they would be out of luck.
Similarly, if a user wanted to add a class or group to a node without replacing
the already assigned classes/groups of a node it was not possible via only
rake. This commit adds a task in the node namespace to add a group or groups to
a node and to add a class or classes to a node. The other two tasks added here
list out the groups or classes already assigned to a node.