Fix descriptions of node tasks to match reality
The descriptions for the node:classes and node:groups tasks are misleading in
that they indicate that the classes and groups of a node can be edited or added
to using the task. The actual tasks will replace, and only replace, the classes
and groups for a node so this commit updates those descriptions to indicate
that.