Remove CHANGELOG file, update version link
This commit removes the CHANGELOG from the top level of Dashboard, which only
contains commit hashes/titles from the previous releases, and is of debateable
utility given the commits aren't particularly useful without git in which case
one could just do `git log`. Currently, the CHANGELOG file in the github repo
forms the target of the version link in the dashboard.  As such, this commit
also updates the dashboard version link to point at the github commit log up to
the version tag specified in VERSION. This allows us to bring the files that
have to be edited and commited to source prior to a Puppet Dashboard release
down to one.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>