Revert "Remove CHANGELOG file, update version link"
The removal of the CHANGELOG was predicated on the notion that no one found it
particularly useful. As it presented a manual release task, its removal was
received wholeheartedly by the release team. However, @dsn pointed out he uses
it for quick-glance change-set recognition, which is certainly a valid use
case. This commit reverts the CHANGELOG removal, in favor of another solution
that removes the VERSION file instead.

This reverts commit 25b5bcaccd9d9e9617e9e19b07721a9f4744c680.