Remove VERSION file, instead using CHANGELOG for version link
This commit removes the VERSION file from dashboard, instead relying on the
CHANGELOG to determine the latest version. This commit retains the behavior of
the version link, which points to the respective CHANGELOG entry in github, but
still removes a manual step in the releasing of Puppet Dashboard. With this
commit, only the CHANGELOG is manually updated, rather than both CHANGELOG and
VERSION.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>