(#13486) Fix DJ doesn't work in a multi-server env
See:

http://projects.puppetlabs.com/issues/13486
http://projects.puppetlabs.com/issues/9409

This commit adds a patch written by Daniel Pittman and posted at
https://gist.github.com/d023814313208dcc1978 but never merged.

It is the case that users want to deploy multiple instances of Puppet
Dashboard and that this minor bug makes what would otherwise be a
decently straightforward multi-instance deployment non-functional. The
fix is simple. It's past due for this capability to be added out-of-box
to product.