(maint) allow running w/ RAILS_ENV=production
Prior to this commit, the rails 3 version of dashboard would not run in
production because of two problems:

 1. bundler didn't know to look for the haml gem outside the :assets
group
 2. the production config explicitly didn't serve static assets (like
images)

This commit adds the haml gem, already a dependency of the existing
haml-rails gem in the :assets group, but outside the :assets group for
availability after assets have been precompiled. This commit also allows
for serving static assets.