Allow rake to execute without full gem bundle
Prior to this commit, all of the gems specified in `Gemfile` were
required for executing any rake task, regardless of whether the command
was prefaced with `bundle exec` or not. This commit uses `rails` as a
proxy to see if we're running in a bootstrap gemset (as is required for
packaging) or in a full, operational gemset.