Set verify_mode explicitly in puppet_https.rb
In Ruby 1.8, Net::HTTP defaulted to SSL::VERIFY_NONE, which meant that the CA
was never verified when submitting the CSR to the puppetmaster/ca. In Ruby 1.9,
the default changed to SSL::VERIFY_PEER, which means that without the CA cert,
the initial CSR cannot be submitted.
The make_ssl_request method in puppet_https.rb takes an authenticate argument,
but the verify_mode was never set to match this parameter.
This commit adds a patch to puppet_https that sets VERIFY_NONE explicitly when
authenticate is set to false in the method, and also sets it explicitly to
VERIFY_PEER when authenticate is set to true. This sets explicitly what was
previously trusted implicitly in the ruby version.