Use the ca_certificate_path when it exists to verify
In the install.rake cert:retrieve method the CA certificate is downloaded and
available to verify future connections to the CA. If the file exists, it should
be used to verify connections when authenticate is true, otherwise the
connections will fail. This commit adds the ca_file to the net http connection
if the ca_certificate_path exists.