(packaging) Use the packaging loader for tasks
The packaging repo now uses an explicit loader which handles loading the
various rake tasks in packaging. This commit updates the puppet-dashboard
Rakefile to use the loader instead of a blind glob of the ext/packaging/tasks
directory. We move this load into the rescue LoadError block because the
packaging repo won't always be there, and define RAKE_ROOT and use it since
its used in more than one place in the Rakefile.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>