Rescue correct psych exception in  ResourceEvent::ValueWrapper
In ruby 1.9, psych raises a `SyntaxError` when it tries to load a string
that is just special characters.  On its own `rescue` only rescues from
a `StandardError` which is a sibling of `SyntaxError` in ruby 1.9 so the
error is not rescued.  Rescuing `Exception` will work in all versions of
ruby.