Fix report format 4 for puppet 4.0 and newer
Very frustratingly puppet slightly changed the report format in puppet 4.0 but did not document the change.  Information about the change can be found at https://tickets.puppetlabs.com/browse/DOCUMENT-436.  In short, this commit takes the hash that is now returned as the tags and converts it into the old style of array.