Make Puppet Dashboard manual accessible
I moved the extracted manual from docs/manual/1.2 -> docs/manual because
I don't think it makes sense to maintain multiple versions. I fixed the
the navigation in the process and linked the manual in the README so
users visiting the Gitlab project have easy access to the docs.

I also removed the "using" chapter because it did not contain any
information, only links to the puppet enterprise documentation.